Added name and rotation attributes to tile and getImagePath in GUI that translates the tile name to a file name.
Added names to tiles in TileStack.
Added eventHandler to add tiles to model when clicked in fieldButtons.
Added update method to GUI that puts pictures in fieldViews according to tabletop TileMap data.
Refactored TUI to use turn from GameData.
Added increment turn Method to tabletop.
TurnCommand is now responsible for incrementing the turn.
Added withTurn Function in GameData to create a GameData with a different turn.